(LedgerStore) Move trait ColumnMetrics and metric-macros to blockstore_metric.rs
